Created and implemented abstract Card.
A Card is a prop in game that is used by human players to deduce the
perpetrator of the murder.

Each Card has a unique ID, a name, and a face.
Names may not be empty or null, and faces may not be null.

Two cards are considered equal iff their IDs are equal.
THe hashcode used is the default one generated by Eclipse.