Removed Displayable as it's not needed.
Displayable was not used by the TextUI and on further discussion we've
decided that it's not to be used on the GUI either. Therefore, we're
removing Displayable from the Model.

Specifically, there are issues with Displayable objects accessing Game
state (e.g. their location) so we're moving to a MVC model instead.