Fixed view hand functionality. Changed image frame to be a dialog.
Controller was passing incorrect names in.
Image frame needed to be changed to a dialog so it could take focus
View hand was only displaying the last item in the players hand.


Known bug:
Only the first player's hand is used for all players.
Problem is likely in the Controller, regarding method calls on
model.getPlayer_____Cards().